Load task list from API
We could have a `Task#get_tasks` method, but I would rather use a
repository pattern approach to keep `Task` lean and respect SRP.

We could also have created a `TaskService` to hold the communication
with the API. The problem with _NounService_ is that they also are an
invitation for large classes with too many responsabilities and could
become just _implementation bags_ and less OO-ish.

So we end with Use Case classes[1][2]. The code is clearer, easy to
maintain and grow. It should have better naming and namespacing though.

The `Task` model is, for now, a `SimpleDelegator` to the
`[Redbooth::Task](https://github.com/teambox/redbooth-ruby/blob/master/lib/redbooth-ruby/task.rb)`.
We do not want the implementation from the gem to leak into our app.

I haven't evaluated the impact in memory allocation and amount of work
for the GC.  Should it become an issue, we could try a lazy enumerable.

[1]: http://devblog.reverb.com/post/70344683203/5-architecture-anti-patterns-and-solutions-for
[2]: http://blog.8thlight.com/uncle-bob/2012/08/13/the-clean-architecture.html