Refresh token and retry last API request
Redbooth `access_token` expires after 30 minutes, with this changes,
instead of redirecting the user to the sign in page and asking to sign
in again, we try to refresh the API token and make it transparent to the
end user. It is a much better solution because we decouple the sessions
of end user and the API client.

Add `APIClient` to encapsulate the redbooth-ruby client (much better!)
and the logic to refresh the token. For now it retries only one time to
avoid possible  infinite loops. In case the retry fails, we raise our
`NotAuthorizedError` and redirects user to sign in.

With the `APIClient` we also gain a better responsabilities separation.
Now `GetTasks` (and future operations) does not need to now anything
about authentication errors.

The user session (controller layer) is inject into the `APIClient`
(business layer). I have mixed feelings about this approach. I feel I'm
pushing an HTTP session down through boundaries it shouldn't cross.
Changing to server-side sessions might be a better solution. I'm not
doing it here to avoid an overkill approach.